Declare gridCols and gridRows as uint32_t
Declare gridCols and gridRows as uint32_t (instead of uint8_t) so that
256 (the maximum legal value) can be represented.

Also make some cleanup changes.

+++ include/avif/avif.h
@@ -889,8 +889,8 @@ enum avifAddImageFlags
 
 AVIF_API avifResult avifEncoderAddImage(avifEncoder * encoder, const avifImage * image, uint64_t durationInTimescales, uint32_t addImageFlags);
 AVIF_API avifResult avifEncoderAddImageGrid(avifEncoder * encoder,
-                                            uint8_t gridCols,
-                                            uint8_t gridRows,
+                                            uint32_t gridCols,
+                                            uint32_t gridRows,
                                             const avifImage * const * cellImages,
                                             uint32_t addImageFlags);
 AVIF_API avifResult avifEncoderFinish(avifEncoder * encoder, avifRWData * output);
